+++ src/test/kotlin/com/polymatus/financialcontrolservice/common/builders/ExpenseRequestBuilder.kt
@@ -1,12 +1,13 @@
 package com.polymatus.financialcontrolservice.common.builders
 
 import com.polymatus.financialcontrolservice.application.web.controllers.dtos.ExpenseRequest
+import java.math.BigDecimal
 
 class ExpenseRequestBuilder {
     var priority: String = "MEDIUM"
     var name: String = "Laptop"
     var category: String = "ELECTRONIC"
-    var price: Double = 15000.0
+    var price: BigDecimal = BigDecimal(15000.0)
     var description: String? = "Dell G15"
     var place: String? = "Amazon"
     var url: String? = "http://amazon.com/laptop-dell-g15"
